feat(network): add receive txs msg check
